Add scripts for additional options on granule status verification
