+++ example/spec/loadTest/ingestLoadTestPassthrough.input.payload.json
@@ -1,0 +1,25 @@
+{
+  "granules": [
+    {
+      "granuleId": "MOD09GQ.A2016358.h13v04.006.2016360104606",
+      "dataType": "MOD09GQ",
+      "version": "006",
+      "files": [
+        {
+          "bucket": "cumulus-test-sandbox-internal",
+          "fileName": "MOD09GQ.A2016358.h13v04.006.2016360104606.hdf",
+          "key": "garbage_path/cumulus-test-data/pdrs/MOD09GQ.A2016358.h13v04.006.2016360104606.hdf",
+          "size": 1098034,
+          "type": "data"
+        },
+        {
+          "bucket": "cumulus-test-sandbox-internal",
+          "fileName": "MOD09GQ.A2016358.h13v04.006.2016360104606.hdf",
+          "key": "garbage_path/cumulus-test-data/pdrs/MOD09GQ.A2016358.h13v04.006.2016360104606.hdf",
+          "size": 21708,
+          "type": "data"
+        }
+      ]
+    }
+  ]
+}
